pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aecd"
version = "0.1.0"
description = "Automated Explanation of Credit Decisions using LLM prompting strategies"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["nlp", "llm", "fraud-detection", "explainability", "prompting", "rouge"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "bitsandbytes>=0.41.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
aecd-run = "scripts.run_experiment:main"

[project.urls]
Homepage = "https://github.com/your-username/aecd"
Issues = "https://github.com/your-username/aecd/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["aecd*"]
